Rename "threaded_dataset" to "internal_dataset"
"internal_dataset" seems to be a better representative name of what
it actually is.
