+++ tomviz/CMakeLists.txt
@@ -498,8 +498,8 @@ set(tomviz_python_modules
   __init__.py
   _internal.py
   operators.py
+  internal_dataset.py
   itkutils.py
-  threaded_dataset.py
   utils.py
   web.py
 )
